% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereqs_checks.R
\name{prereq_server}
\alias{prereq_server}
\title{Returns an empty shiny server variable if the prereqs checks fail}
\usage{
prereq_server(server)
}
\arguments{
\item{server}{A shiny server object}
}
\value{
either the input or an empty server object
}
\description{
Tests the defined pre-requisites and either returns the input
(on success) or an empty server object if any of the pre-reqs
fail
}
