% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereqs_success.R
\name{prereq_success}
\alias{prereq_success}
\title{Runs the pre-requisites tests}
\usage{
prereq_success()
}
\value{
A list containing the test status and a vector of failed tests
}
\description{
Runs the pre-requisites tests
}
