% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereqs_checks.R
\name{prereq_ui}
\alias{prereq_ui}
\title{Returns an error display app if the prereqs checks fail}
\usage{
prereq_ui(ui)
}
\arguments{
\item{ui}{A shiny ui object}
}
\value{
either the input or an error report ui object
}
\description{
Tests the defined pre-requisites and either returns the input
(on success) or a error report ui object if any of the pre-reqs
fail. It effectively replaces the input UI in the event of a
pre-req check failure.
}
