% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_prereqs.R
\name{use_prereqs}
\alias{use_prereqs}
\title{writes a skeleton yaml config file for shinyprereqs}
\usage{
use_prereqs(path = ".")
}
\arguments{
\item{path}{A path to the directory containing your app}
}
\value{
Invisibly returns the path
}
\description{
Creates the necessary configuration file used to specify
the pre-requisites. This function creates a skeleton
file that you can edit to suit your project.
}
